refactor(shared): extract openrouter metadata module
